Added logitech webcam parameters and a main using webcam
